requirements.txt
Flask>=2.2
python-chess>=1.999
